fbdev: msm: somc_panel: Complete restructuration of color correction
The color correction code has been cleaned up and separated from
the panel driver itself.
Now it is called somc_panel Color Manager and lives in its own
file in the somc_panel directory. Also the specific_pdata struct
has been cleaned up and now the Color Manager specific variables
live in their Color Manager's own struct.
