+++ drivers/video/fbdev/msm/somc_panel/Makefile
@@ -6,6 +6,7 @@ obj-$(CONFIG_FB_MSM_MDSS_SPECIFIC_PANEL)	+= common.o
 obj-$(CONFIG_FB_MSM_MDSS_SPECIFIC_PANEL)	+= panel_detect.o
 obj-$(CONFIG_FB_MSM_MDSS_SPECIFIC_PANEL)	+= panel_polling.o
 obj-$(CONFIG_FB_MSM_MDSS_SPECIFIC_PANEL)	+= panel_fps_manager.o
+obj-$(CONFIG_FB_MSM_MDSS_SPECIFIC_PANEL)	+= panel_color_manager.o
 obj-$(CONFIG_FB_MSM_MDSS_SPECIFIC_PANEL)	+= panel_driver.o
 obj-$(CONFIG_SOMC_PANEL_LABIBB)			+= panel_regulators.o
 obj-$(CONFIG_SOMC_PANEL_LEGACY)			+= panel_legacy.o
